u9fs: Use explicitly sized types in most places.

+++ authp9any.c
@@ -79,7 +79,7 @@ struct Authenticator
 {
 	char	num;			/* replay protection */
 	char	chal[CHALLEN];
-	ulong	id;			/* authenticator id, ++'d with each auth */
+	uint32_t id;			/* authenticator id, ++'d with each auth */
 };
 #define	AUTHENTLEN	(CHALLEN+4+1)
 
@@ -154,16 +154,16 @@ static int
 convTR2M(Ticketreq *f, char *ap)
 {
 	int n;
-	uchar *p;
+	uint8_t *p;
 
-	p = (uchar*)ap;
+	p = (uint8_t*)ap;
 	CHAR(type);
 	STRING(authid, NAMELEN);
 	STRING(authdom, DOMLEN);
 	STRING(chal, CHALLEN);
 	STRING(hostid, NAMELEN);
 	STRING(uid, NAMELEN);
-	n = p - (uchar*)ap;
+	n = p - (uint8_t*)ap;
 	return n;
 }
 
@@ -171,13 +171,13 @@ int
 convA2M(Authenticator *f, char *ap, char *key)
 {
 	int n;
-	uchar *p;
+	uint8_t *p;
 
-	p = (uchar*)ap;
+	p = (uint8_t*)ap;
 	CHAR(num);
 	STRING(chal, CHALLEN);
 	LONG(id);
-	n = p - (uchar*)ap;
+	n = p - (uint8_t*)ap;
 	if(key)
 		encrypt9p(key, ap, n);
 	return n;
@@ -198,11 +198,11 @@ convA2M(Authenticator *f, char *ap, char *key)
 void
 convM2A(char *ap, Authenticator *f, char *key)
 {
-	uchar *p;
+	uint8_t *p;
 
 	if(key)
 		decrypt9p(key, ap, AUTHENTLEN);
-	p = (uchar*)ap;
+	p = (uint8_t*)ap;
 	CHAR(num);
 	STRING(chal, CHALLEN);
 	LONG(id);
@@ -212,11 +212,11 @@ convM2A(char *ap, Authenticator *f, char *key)
 void
 convM2T(char *ap, Ticket *f, char *key)
 {
-	uchar *p;
+	uint8_t *p;
 
 	if(key)
 		decrypt9p(key, ap, TICKETLEN);
-	p = (uchar*)ap;
+	p = (uint8_t*)ap;
 	CHAR(num);
 	STRING(chal, CHALLEN);
 	STRING(cuid, NAMELEN);
@@ -236,7 +236,7 @@ convM2T(char *ap, Ticket *f, char *key)
 static int
 passtokey(char *key, char *p)
 {
-	uchar buf[NAMELEN], *t;
+	uint8_t buf[NAMELEN], *t;
 	int i, n;
 
 	n = strlen(p);
@@ -437,7 +437,7 @@ p9anywrite(Fcall *rx, Fcall *tx)
 		sp->tr.type = AuthTreq;
 		safecpy(sp->tr.authid, authid, sizeof(sp->tr.authid));
 		safecpy(sp->tr.authdom, authdom, sizeof(sp->tr.authdom));
-		randombytes((uchar *)sp->tr.chal, CHALLEN);
+		randombytes((uint8_t *)sp->tr.chal, CHALLEN);
 		safecpy(sp->tr.hostid, "", sizeof(sp->tr.hostid));
 		safecpy(sp->tr.uid, "", sizeof(sp->tr.uid));
 		tx->count = rx->count;
